debugged method for asserting particles equality and test io

+++ particles/tests/mod_particle_io_mpi_test.f90
@@ -15,7 +15,9 @@ type(particle_sim)          :: sim_particles
 character(len=28),parameter :: test_filename="test_particle_io_mpi_hdf5.h5"
 !> the number of particle groups is set equal to the number
 !> of particle types for testing all of them
-integer,parameter :: n_groups=8
+!> particle_gc_Qin is commented because the I/O for particle_gc_Qin
+!> has not been implemented yet
+integer,parameter :: n_groups=7!8
 integer,parameter :: n_particles=5 !< N# of particles per group per task
 real*8,parameter :: tol_real8=1.d-15
 !> intervals for random number generation
@@ -26,7 +28,6 @@ integer,dimension(2),parameter   :: i_life_interval=(/1,1000000/)
 real*8,dimension(2),parameter    :: sim_time_interval=(/0.d0,1.d3/)
 real*8,dimension(2),parameter    :: t_birth_interval=(/0.d0,3.45d4/)
 real*8,dimension(2),parameter    :: st_interval=(/0.d0,1.d0/)
-real*8,dimension(2),parameter    :: dt_interval=(/1.d-12,1.d-2/)
 real*8,dimension(2),parameter    :: mass_interval=(/5.485d-4,124.d0/)
 real*8,dimension(2),parameter    :: v_interval=(/-6.75d3,8.45d3/)
 real*8,dimension(2),parameter    :: Ekin_interval=(/0.d0,1.d7/)
@@ -97,31 +98,31 @@ subroutine setup(rank,n_tasks,ifail)
   if(rank.eq.0) then
     call gnu_rng_interval(sim_time_interval,sim_particles%time)
   endif
-  call MPI_Bcast(sim_particles%time,1,MPI_REAL8,0,MPI_COMM_WORLD)
+  call MPI_Bcast(sim_particles%time,1,MPI_REAL8,0,MPI_COMM_WORLD,ifail)
 
   !> allocate particle lists for different particle types
   allocate(particle_fieldline::sim_particles%groups(1)%particles(n_particles))
   allocate(particle_gc::sim_particles%groups(2)%particles(n_particles))
   allocate(particle_gc_vpar::sim_particles%groups(3)%particles(n_particles))
-  allocate(particle_gc_Qin::sim_particles%groups(4)%particles(n_particles))
-  allocate(particle_kinetic::sim_particles%groups(5)%particles(n_particles))
-  allocate(particle_kinetic_leapfrog::sim_particles%groups(6)%particles(n_particles))
-  allocate(particle_kinetic_relativistic::sim_particles%groups(7)%particles(n_particles))
-  allocate(particle_gc_relativistic::sim_particles%groups(8)%particles(n_particles))
+  allocate(particle_kinetic::sim_particles%groups(4)%particles(n_particles))
+  allocate(particle_kinetic_leapfrog::sim_particles%groups(5)%particles(n_particles))
+  allocate(particle_kinetic_relativistic::sim_particles%groups(6)%particles(n_particles))
+  allocate(particle_gc_relativistic::sim_particles%groups(7)%particles(n_particles))
+  !allocate(particle_gc_Qin::sim_particles%groups(8)%particles(n_particles)) !< IO not implemented
 
   !> fill-up the groupd and particle base variables
-  call fill_sim_groups(rank)
+  call fill_sim_groups(rank,ifail)
   call fill_particle_base(rank)
 
   !> fill up variables for each species
   call fill_particle_fieldline(rank,sim_particles%groups(1)%particles)
   call fill_particle_gc(rank,sim_particles%groups(2)%particles)
   call fill_particle_gc_vpar(rank,sim_particles%groups(3)%particles)
-  call fill_particle_gc_Qin(rank,sim_particles%groups(4)%particles)
-  call fill_particle_kinetic(rank,sim_particles%groups(5)%particles)
-  call fill_particle_kinetic_leapfrog(rank,sim_particles%groups(6)%particles)
-  call fill_particle_kinetic_relativistic(rank,sim_particles%groups(7)%particles)
-  call fill_particle_gc_relativistic(rank,sim_particles%groups(8)%particles)
+  call fill_particle_kinetic(rank,sim_particles%groups(4)%particles)
+  call fill_particle_kinetic_leapfrog(rank,sim_particles%groups(5)%particles)
+  call fill_particle_kinetic_relativistic(rank,sim_particles%groups(6)%particles)
+  call fill_particle_gc_relativistic(rank,sim_particles%groups(7)%particles)
+  !call fill_particle_gc_Qin(rank,sim_particles%groups(8)%particles) !< I/O not implemented 
 
   !> write default simulation in file and read it in new simulation
   call write_simulation_hdf5(sim_particles,trim(test_filename))
@@ -142,8 +143,9 @@ subroutine teardown(rank,n_tasks,ifail)
   !> inputs-outputs
   integer,intent(inout) :: ifail
 
+  call MPI_Barrier(MPI_COMM_WORLD,ifail)
   !> remove test file
-  call system("rm "//test_filename)
+  if(rank.eq.0) call system("rm "//test_filename)
 end subroutine teardown
 !> Tests ------------------------------------------------
 !> procedure for testing the particle io
@@ -159,6 +161,7 @@ subroutine test_particle_mpi_io(rank,n_tasks,ifail)
   !> variables
   type(particle_sim) :: sim_particles_new
   integer :: ii
+  real*8 :: comp_real8_1,comp_real8_2
 
   !> initialize the new particle simulation (requires jorek inputfile)
   call sim_particles_new%initialize(n_groups,.false.,rank,n_tasks)
@@ -166,39 +169,79 @@ subroutine test_particle_mpi_io(rank,n_tasks,ifail)
   !> read default simulation from file and store in new sim
   call read_simulation_hdf5(sim_particles_new,trim(test_filename))
 
+  !> compu variables which are not read from hdf5
+  call copy_sim_fieldline_B_hat_prev(sim_particles,sim_particles_new)
+
   !> check simulation 
   call assert_equals(sim_particles_new%time,sim_particles%time,tol_real8,&
   "Error writing/reading particle simulation: time mismatch!")
+
   !> check groups
   do ii=1,n_groups
     call assert_equals(sim_particles_new%groups(ii)%mass,sim_particles%groups(ii)%mass,&
     tol_real8,"Error writing/reading particle simulation: mass mismatch!")
-    call assert_equals(sim_particles_new%groups(ii)%dt,sim_particles%groups(ii)%dt,&
-    tol_real8,"Error writing/reading particle simulation: dt mismatch!")
     call assert_equal_particle(n_particles,sim_particles_new%groups(ii)%particles,&
     sim_particles%groups(ii)%particles)
   enddo
-
 end subroutine test_particle_mpi_io
 
 !> Tools ------------------------------------------------
+!> copy fieldlines B_hat between two simulations 
+!> used for IO because it is not stored in hdf5
+subroutine copy_sim_fieldline_B_hat_prev(sim_in,sim_out)
+  use mod_particle_types, only: particle_fieldline
+  use mod_particle_sim,   only: particle_sim
+  implicit none
+  type(particle_sim),intent(in)    :: sim_in
+  type(particle_sim),intent(inout) :: sim_out
+  integer :: ii,jj
+  !$omp parallel default(private) shared(sim_in,sim_out)
+  do ii=1,n_groups
+    !$omp do
+    do jj=1,n_particles
+      select type (p_out=>sim_out%groups(ii)%particles(jj))
+      type is (particle_fieldline)
+        select type (p_in=>sim_in%groups(ii)%particles(jj))
+          type is (particle_fieldline)
+          p_out%B_hat_prev = p_in%B_hat_prev
+        end select
+      end select
+    enddo
+    !$omp end do
+  enddo
+  !$omp end parallel
+
+end subroutine copy_sim_fieldline_B_hat_prev
+
 !> generate random values for filling the groups type
-subroutine fill_sim_groups(rank)
+!> we do not create random mass for all groups because
+!> it seems that only rank 0 is saved in hdf5
+subroutine fill_sim_groups(rank,ifail)
   use mod_gnu_rng, only: gnu_rng_interval
   use mod_gnu_rng, only: set_seed_sys_time
   implicit none
   !> inputs
   integer,intent(in) :: rank
+  !> input-outputs
+  integer,intent(inout) :: ifail
   !> variables
   integer :: ii
   real*8 :: rn_real
-  call set_seed_sys_time(rng_seed_interval,rank)
-  do ii=1,n_groups
-    call gnu_rng_interval(mass_interval,rn_real)
-    sim_particles%groups(ii)%mass = rn_real
-    call gnu_rng_interval(dt_interval,rn_real)
-    sim_particles%groups(ii)%dt = rn_real
-  enddo
+  real*8,dimension(n_groups) :: val_to_bcst
+  if(rank.eq.0) then
+    call set_seed_sys_time(rng_seed_interval,rank)
+    do ii=1,n_groups
+      call gnu_rng_interval(mass_interval,rn_real)
+      sim_particles%groups(ii)%mass = rn_real
+      val_to_bcst(ii) = sim_particles%groups(ii)%mass
+    enddo
+  endif
+  call MPI_Bcast(val_to_bcst,n_groups,MPI_REAL8,0,MPI_COMM_WORLD,ifail)
+  if(rank.ne.0) then
+    do ii=1,n_groups
+      sim_particles%groups(ii)%mass = val_to_bcst(ii)
+    enddo
+  endif
 end subroutine fill_sim_groups
 
 !> generate random values for filling the particle base type
